feat(tests): cover ModExp gas before EIP-2565
The gas measurement harness lived in the EIP-7883 directory and every
test using it started at Berlin, so the EIP-198 schedule that prices the
precompile from Byzantium until Berlin had no coverage: the only ModExp
tests reaching those forks call with all the gas available.

Move the harness down to the EIP-198 directory, which already owns
`ModExpInput` and is imported by the later repricing EIPs, and select
the gas formula through a `modexp_spec` fixture that a repricing
directory overrides. The formulas move with it: `Spec198` is new,
`Spec2565` is the former `Spec` body, and EIP-7883 keeps overriding it.

Two fixes were needed to reach Byzantium. The harness subtracted
`gas_costs().WARM_ACCESS` as the cost of the call, but that field holds
the EIP-2929 value on every fork, so before Berlin it under-counted the
flat EIP-150 charge by 600 gas. And `get_exponent_head()` left-padded an
exponent whose input is truncated, while reading past the end of the
input yields zeros on the right; the EIP-2565 minimum of 200 gas had
hidden the difference, which is observable under EIP-198 where a call
can be free.

With that in place, the legacy vectors and the exact-gas boundary test
now run from Byzantium, and a new test pins each branch of the EIP-198
multiplication complexity formula at its exact cost and one gas less.
